fix: unit tests, removed unused functions
